test: 당첨 내역 계산 기능 책임 이동으로 인한 테스트 리팩토링
